Added cron_shared partial. Fixed provider cron spec failure. Moved cron and cron resource into the lib/chef/resource/cron directory.
