+++ lib/chef/resources.rb
@@ -38,9 +38,9 @@ require_relative "resource/chocolatey_config"
 require_relative "resource/chocolatey_feature"
 require_relative "resource/chocolatey_package"
 require_relative "resource/chocolatey_source"
-require_relative "resource/cron"
+require_relative "resource/cron/cron"
 require_relative "resource/cron_access"
-require_relative "resource/cron_d"
+require_relative "resource/cron/cron_d"
 require_relative "resource/csh"
 require_relative "resource/directory"
 require_relative "resource/dmg_package"
